+++ src/from-promise.ts
@@ -1,4 +1,4 @@
-import {IObservableValue, observable, action} from "mobx";
+import {IObservableValue, extendShallowObservable, action} from "mobx";
 import {IDENTITY, deprecated, invariant} from "./utils";
 
 export type PromiseState = "pending" | "fulfilled" | "rejected";
@@ -8,9 +8,9 @@ export const FULFILLED = "fulfilled";
 export const REJECTED = "rejected";
 
 export type IBasePromiseBasedObservable<T> = {
-    readonly promise: PromiseLike<T>;
+    isPromiseBasedObservable: true,
     case<U>(handlers: {pending?: () => U, fulfilled?: (t: T) => U, rejected?: (e: any) => U}): U;
-}
+} & PromiseLike<T>
 
 export type IPendingPromise = {
     readonly state: "pending";
@@ -29,30 +29,56 @@ export type IRejectedPromise = {
 
 export type IPromiseBasedObservable<T> = IBasePromiseBasedObservable<T> & (IPendingPromise | IFulfilledPromise<T> | IRejectedPromise)
 
-class PromiseBasedObservable<T> {
-    @observable.ref value: T = undefined;
-    @observable.ref state: PromiseState = PENDING;
+function caseImpl<U, T>(handlers: {pending?: () => U, fulfilled?: (t: T) => U, rejected?: (e: any) => U}): U {
+    switch (this.state) {
+        case PENDING: return handlers.pending && handlers.pending();
+        case REJECTED: return handlers.rejected && handlers.rejected(this.value);
+        case FULFILLED: return handlers.fulfilled && handlers.fulfilled(this.value);
+    }
+}
 
-    constructor(public promise: PromiseLike<T>) {
-        promise.then(
-            action("observableFromPromise-resolve", (value: T) => {
-                this.value = value;
-                this.state = FULFILLED;
-            }),
-            action("observableFromPromise-reject", (reason: any) => {
-                this.value = reason;
-                this.state = REJECTED;
-            })
-        );
+function createObservablePromise(origPromise: any) {
+    invariant(arguments.length === 1, "fromPromise expects exactly one argument");
+    invariant(
+        typeof origPromise === "function" ||
+            (typeof origPromise === "object" && origPromise && typeof origPromise.then === "function"),
+        "Please pass a promise or function to fromPromise"
+    );
+
+    if (typeof origPromise === "function") {
+        // If it is a (reject, resolve function, wrap it)
+        origPromise = new Promise(origPromise as any);
     }
 
-    public case<U>(handlers: {pending?: () => U, fulfilled?: (t: T) => U, rejected?: (e: any) => U}): U {
-        switch (this.state) {
-            case PENDING: return handlers.pending && handlers.pending();
-            case REJECTED: return handlers.rejected && handlers.rejected(this.value);
-            case FULFILLED: return handlers.fulfilled && handlers.fulfilled(this.value);
+    const promise = new Promise((resolve, reject) => {
+        origPromise.then(action("observableFromPromise-resolve", (value: any) => {
+            promise.value = value;
+            promise.state = FULFILLED;
+            resolve(value);
+        }),
+        action("observableFromPromise-reject", (reason: any) => {
+            promise.value = reason;
+            promise.state = REJECTED;
+            reject(reason);
+        }));
+    }) as any;
+
+    promise.isPromiseBasedObservable = true;
+    promise.case = caseImpl;
+    extendShallowObservable(promise, {
+        value: undefined,
+        state: PENDING
+    });
+
+    // TODO: remove in next major
+    Object.defineProperty(promise, "promise", {
+        get() {
+            deprecated("fromPromise().promise is deprecated. fromPromise now directly returns a promise");
+            return origPromise;
         }
-    }
+    });
+
+    return promise;
 }
 
 /**
@@ -64,7 +90,19 @@ class PromiseBasedObservable<T> {
  * and the following method:
  * - `case({fulfilled, rejected, pending})`: maps over the result using the provided handlers, or returns `undefined` if a handler isn't available for the current promise state.
  *
+ *
+ * Note that the status strings are available as constants:
+ * `mobxUtils.PENDING`, `mobxUtils.REJECTED`, `mobxUtil.FULFILLED`
+ *
+ * Observable promises can be created immediatly in a certain state using
+ * `fromPromise.reject(reason)` or `fromPromise.resolve(value?)`.
+ * The mean advantagate of `fromPromise.resolve(value)` over `fromPromise(Promise.resolve(value))` is that the first _synchronously_ starts in the desired state.
+ *
+ * It is possible to directly create a promise using a resolve, reject function:
+ * `fromPromise((resolve, reject) => setTimeout(() => resolve(true), 1000))`
+ *
  * @example
+ * ```javascript
  * const fetchResult = fromPromise(fetch("http://someurl"))
  *
  * // combine with when..
@@ -92,13 +130,7 @@ class PromiseBasedObservable<T> {
  *     rejected:  error => <div>Ooops.. {error}</div>
  *     fulfilled: value => <div>Gotcha: {value}</div>
  *   }))
- *
- * Note that the status strings are available as constants:
- * `mobxUtils.PENDING`, `mobxUtils.REJECTED`, `mobxUtil.FULFILLED`
- *
- * Observable promises can be created immediatly in a certain state using
- * `fromPromise.reject(reason)` or `fromPromise.resolve(value?)`.
- * The mean advantagate of `fromPromise.resolve(value)` over `fromPromise(Promise.resolve(value))` is that the first _synchronously_ starts in the desired state.
+ * ```
  *
  * @param {IThenable<T>} promise The promise which will be observed
  * @returns {IPromiseBasedObservable<T>}
@@ -107,11 +139,7 @@ export const fromPromise: {
     <T>(promise: PromiseLike<T>): IPromiseBasedObservable<T>;
     reject<T>(reason: any): IRejectedPromise & IBasePromiseBasedObservable<T>;
     resolve<T>(value?: T): IFulfilledPromise<T> & IBasePromiseBasedObservable<T>;
-} = function(promise: any) {
-    invariant(arguments.length === 1, "fromPromise expects exactly one argument");
-    invariant(typeof promise === "object" && promise && typeof promise.then === "function", "Please pass a promise to fromPromise");
-    return new PromiseBasedObservable(promise) as any;
-} as any;
+} = createObservablePromise as any;
 
 fromPromise.reject = action("fromPromise.reject", function(reason: any) {
     const p: any = fromPromise(Promise.reject(reason));
@@ -133,5 +161,5 @@ fromPromise.resolve = action("fromPromise.resolve", function(value: any = undefi
  * @returns {boolean}
  */
 export function isPromiseBasedObservable(value: any): value is IPromiseBasedObservable<any> {
-    return value instanceof PromiseBasedObservable;
+    return value && value.isPromiseBasedObservable === true;
 }